Implemented base for the interstection algo

+++ Lab1/Lab1/Matrix.cs
@@ -8,11 +8,11 @@ namespace Lab1
 {
     internal class Matrix
     {
-        private readonly double[][] values;
+        private readonly float[][] values;
         public readonly int N;
         public readonly int M;
 
-        public Matrix(double[][] values)
+        public Matrix(float[][] values)
         {
             if (values == null)
             {
@@ -20,10 +20,10 @@ namespace Lab1
             }
             N = values.Length;
             M = values[0].Length;
-            this.values = new double[N][];
+            this.values = new float[N][];
             for (int i = 0; i < N; i++)
             {
-                this.values[i] = new double[M];
+                this.values[i] = new float[M];
                 for (int j = 0; j < M; j++)
                 {
                     this.values[i][j] = values[i][j];
@@ -31,7 +31,7 @@ namespace Lab1
             }
         }
 
-        public Matrix(double[] values, int n, int m)
+        public Matrix(float[] values, int n, int m)
         {
             if (values == null)
             {
@@ -39,10 +39,10 @@ namespace Lab1
             }
             N = n;
             M = m;
-            this.values = new double[N][];
+            this.values = new float[N][];
             for (int i = 0; i < N; i++)
             {
-                this.values[i] = new double[M];
+                this.values[i] = new float[M];
                 for (int j = 0; j < M; j++)
                 {
                     this.values[i][j] = values[i * N + j];
@@ -50,16 +50,16 @@ namespace Lab1
             }
         }
 
-        public double[][] Values() => this.values;
+        public float[][] Values() => this.values;
 
         public static Matrix operator *(Matrix matrix1, Matrix matrix2)
         {
-            double[][] vals1 = matrix1.Values();
-            double[][] vals2 = matrix2.Values();
-            double[][] res = new double[matrix1.N][];
+            float[][] vals1 = matrix1.Values();
+            float[][] vals2 = matrix2.Values();
+            float[][] res = new float[matrix1.N][];
             for (int i = 0; i < matrix1.N; i++)
             {
-                res[i] = new double[matrix2.M];
+                res[i] = new float[matrix2.M];
                 for (int j = 0; j < matrix2.M; j++)
                 {
                     res[i][j] = 0;
